Rename prospect.data -> prospect.observation. Also slight change to instrumental smoothing.
