--- prospect/data/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .observation import Photometry, Spectrum, from_oldstyle
-
-__all__ = ["Photometry", "Spectrum",
-           "from_oldstyle"]
