+++ prospect/observation/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from .observation import Photometry, Spectrum, Lines, from_oldstyle
+
+__all__ = ["Photometry", "Spectrum", "Lines",
+           "from_oldstyle"]
